CRUCIAL CHANGE: Entities data are now read of the filesystem as normal files.
There are 2 main reasons for it:

1) PERFORMANCE.
2) Concurrency.

About 2, there were various elements in the entities get pipeline that relied on streams, I still do not know how to combine streans with old callbacks (IE on("end") function being called) and make the express response (res) waiting for that callback. So I took the chance to change that system that looked fragile anyway and dumped entities data out of MongoDB gridFS straight into the normal filesystem. That way we can also take advantage of a future CDN.